Commit to messaging and Shared library change
