+++ src/AuthoBson.Messaging/AuthoBson.Messaging.csproj
@@ -13,8 +13,14 @@
   </PropertyGroup>
 
     <ItemGroup>
-      <Folder Include="Forms" />
-      <Folder Include="ViewModels" />
+      <Compile Remove="Forms\**" />
+      <Compile Remove="ViewModels\**" />
+      <Content Remove="Forms\**" />
+      <Content Remove="ViewModels\**" />
+      <EmbeddedResource Remove="Forms\**" />
+      <EmbeddedResource Remove="ViewModels\**" />
+      <None Remove="Forms\**" />
+      <None Remove="ViewModels\**" />
     </ItemGroup>
 
     <ItemGroup>
