+++ src/AuthoBson.Messaging/Data/Models/Message.cs
@@ -3,6 +3,7 @@ using System.Collections;
 using System.Text;
 using System.IO;
 using System.Linq;
+using AuthoBson.Shared;
 using AuthoBson.Shared.Data.Models;
 using MongoDB.Bson;
 using MongoDB.Bson.Serialization.Attributes;
@@ -61,9 +62,19 @@ namespace AuthoBson.Messaging.Data.Models {
 
         [BsonConstructor("SenderId", "ReceiverId", "Header", "Body")]
         public Message(string SenderId, string ReceiverId, string Header, string Body) {
-            this.Id = ObjectId.GenerateNewId().ToString();
+            Id = ObjectId.GenerateNewId().ToString();
             this.SenderId = SenderId;
             this.ReceiverId = ReceiverId;
+            this.Header = Header;
+            this.Body = Body;
+        }
+
+        [BsonConstructor("Sender", "Receiver", "Header", "Body")]
+        public Message([Messaging] ModelBase Sender, [Messaging(true)] ModelBase Receiver, string Header, string Body)
+        {
+            Id = ObjectId.GenerateNewId().ToString();
+            SenderId = Sender.Id;
+            ReceiverId = Receiver.Id;
             this.Header = Header;
             this.Body = Body;
         }
