+++ src/AuthoBson/AuthoBson.csproj.user
@@ -4,6 +4,6 @@
     <DebuggerFlavor>ProjectDebugger</DebuggerFlavor>
   </PropertyGroup>
   <PropertyGroup>
-    <ActiveDebugProfile>IIS Express</ActiveDebugProfile>
+    <ActiveDebugProfile>AuthoBson</ActiveDebugProfile>
   </PropertyGroup>
 </Project>
